# Optimize FindMatchLength for x64