--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -43,21 +43,62 @@ jobs:
     - name: Test
       run: |
         export COMPlus_Enable${{ matrix.disable }}=0 && \
-        dotnet test --no-build -f ${{ matrix.framework }} --configuration Release --verbosity normal --logger "trx;LogFileName=results.trx" && \
-        dotnet test --runtime linux-x86 -f ${{ matrix.framework }} --configuration Release --verbosity normal --logger "trx;LogFileName=results-x86.trx"
+        dotnet test --no-build -f ${{ matrix.framework }} --configuration Release --verbosity normal --logger "trx;LogFileName=results.trx"
 
     - name: Test Report
       uses: dorny/test-reporter@v1
       if: success() || failure() # run this step even if previous step failed
       with:
         name: Unit Tests ${{ matrix.framework }} ${{ matrix.disable }}
-        path: "**/results*.trx"
+        path: "**/results.trx"
+        reporter: dotnet-trx
+
+  test-windows:
+
+    runs-on: windows-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Setup .NET 7
+      uses: actions/setup-dotnet@v3
+      with:
+        dotnet-version: '7.0.x'
+    # Cache packages for faster subsequent runs
+    - uses: actions/cache@v3
+      with:
+        path: ~/.nuget/packages
+        key: ${{ runner.os }}-nuget-${{ hashFiles('**/*.csproj') }}
+        restore-keys: |
+          ${{ runner.os }}-nuget-
+
+    - name: Install dependencies
+      run: dotnet restore
+    - name: Test
+      run: |-
+        dotnet test --runtime win10-x64 -f net48 --configuration Release --verbosity normal --logger "trx;LogFileName=results-x64.trx"
+        dotnet test --runtime win10-x86 -f net48 --configuration Release --verbosity normal --logger "trx;LogFileName=results-x86.trx"
+
+    - name: Test Report x64
+      uses: dorny/test-reporter@v1
+      if: success() || failure() # run this step even if previous step failed
+      with:
+        name: Unit Tests Windows x64
+        path: "**/results-x64.trx"
+        reporter: dotnet-trx
+    - name: Test Report x86
+      uses: dorny/test-reporter@v1
+      if: success() || failure() # run this step even if previous step failed
+      with:
+        name: Unit Tests Windows x86
+        path: "**/results-x86.trx"
         reporter: dotnet-trx
 
   publish:
 
     runs-on: ubuntu-latest
-    needs: test
+    needs:
+    - test
+    - test-windows
     if: ${{ startsWith(github.ref, 'refs/tags/release/') }}
 
     steps: