--- a/Snappier.Tests/Snappier.Tests.csproj
+++ b/Snappier.Tests/Snappier.Tests.csproj
@@ -8,6 +8,9 @@
     <LangVersion>11</LangVersion>
     <SignAssembly>true</SignAssembly>
     <AssemblyOriginatorKeyFile>..\Snappier.snk</AssemblyOriginatorKeyFile>
+
+    <!-- Prevents a warning when passing runtime to tests on the commandline -->
+    <SelfContained>false</SelfContained>
   </PropertyGroup>
 
   <ItemGroup>