… Snappier.Benchmarks/Log2FloorHelper.cs