Motivation
----------
On little endian 64-bit CPUs, such as Intel/AMD x64, we can optimize FindMatchLength to work in 8 byte chunks instead of 4 bytes.

Modifications
-------------
Add an optimized path that is chosen by JIT based on CPU endianness and pointer size.

Results
-------
This benchmark shows the improvement on x64 for compressing HTML data.

BenchmarkDotNet=v0.13.4, OS=Windows 11 (10.0.22000.1641/21H2) Intel Core i7-10850H CPU 2.70GHz, 1 CPU, 12 logical and 6 physical cores .NET SDK=7.0.200
  [Host]     : .NET 7.0.3 (7.0.323.6910), X64 RyuJIT AVX2
  Job-XYFPAG : .NET 7.0.3 (7.0.323.6910), X64 RyuJIT AVX2
  Job-KJJRVR : .NET 6.0.14 (6.0.1423.7309), X64 RyuJIT AVX2
  Job-RNFVAW : .NET 7.0.3 (7.0.323.6910), X64 RyuJIT AVX2
  Job-EUIBUW : .NET Framework 4.8 (4.8.4614.0), X64 RyuJIT VectorSize=256

|     Method |            Runtime | PGO |     Mean |   Error |  StdDev | Ratio | Rank |
|----------- |------------------- |---- |---------:|--------:|--------:|------:|-----:|
| Match32Bit |           .NET 7.0 |   Y | 119.5 us | 0.53 us | 0.44 us |  1.00 |    2 |
| Match64Bit |           .NET 7.0 |   Y | 104.9 us | 0.12 us | 0.10 us |  0.88 |    1 |
|            |                    |     |          |         |         |       |      |
| Match32Bit |           .NET 6.0 |   N | 120.5 us | 0.23 us | 0.19 us |  1.00 |    2 |
| Match64Bit |           .NET 6.0 |   N | 114.2 us | 0.39 us | 0.33 us |  0.95 |    1 |
|            |                    |     |          |         |         |       |      |
| Match32Bit |           .NET 7.0 |   N | 124.6 us | 1.91 us | 1.49 us |  1.00 |    2 |
| Match64Bit |           .NET 7.0 |   N | 108.0 us | 0.47 us | 0.37 us |  0.87 |    1 |
|            |                    |     |          |         |         |       |      |
| Match32Bit | .NET Framework 4.8 |   N | 206.1 us | 0.27 us | 0.25 us |  1.00 |    2 |
| Match64Bit | .NET Framework 4.8 |   N | 190.8 us | 0.12 us | 0.10 us |  0.93 |    1 |